Employee_list/requirements.txt
certifi==2019.6.16
Django==2.2.3
djangorestframework==3.10.2
Pillow==6.1.0
pipenv==2018.11.26
psycopg2==2.8.3
pytz==2019.1
sqlparse==0.3.0
virtualenv==16.6.2
virtualenv-clone==0.5.3
